+++ src/Symfony/Component/Notifier/Bridge/Mercure/Tests/MercureOptionsTest.php
@@ -24,16 +24,13 @@ final class MercureOptionsTest extends TestCase
             'id' => null,
             'type' => null,
             'retry' => null,
-            'body' => null,
-            'icon' => null,
-            'tag' => null,
-            'renotify' => false,
+            'content' => null
         ]);
     }
 
     public function testConstructWithParameters()
     {
-        $options = (new MercureOptions('/topic/1', true, 'id', 'type', 1));
+        $options = (new MercureOptions('/topic/1', true, 'id', 'type', 1, null));
 
         $this->assertSame($options->toArray(), [
             'topics' => ['/topic/1'],
@@ -41,10 +38,7 @@ final class MercureOptionsTest extends TestCase
             'id' => 'id',
             'type' => 'type',
             'retry' => 1,
-            'body' => null,
-            'icon' => null,
-            'tag' => null,
-            'renotify' => false,
+            'content' => null
         ]);
     }